% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_calibrate_dmat.R
\name{aux_calibrate_dmat}
\alias{aux_calibrate_dmat}
\title{Calibrate distance matrix}
\usage{
aux_calibrate_dmat(dmat)
}
\arguments{
\item{dmat}{Distance matrix}
}
\value{
Calibrated distance matrix
}
\description{
Calibrate distance matrix to better suit t-SNE
}
