% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_get_dist_super.R
\name{aux_get_dist_super}
\alias{aux_get_dist_super}
\title{Generate a supervised distance matrix with a fixed shift value}
\usage{
aux_get_dist_super(mat = NULL, dmat = NULL, class.labels, s = 10)
}
\arguments{
\item{mat}{A data frame or matrix to be processed. Rows are variables and columns are samples.}

\item{dmat}{A precomputed distance matrix. Used as an alternative input to \code{mat}.}

\item{class.labels}{A vector of class labels}

\item{s}{Shift value added to distances between samples from different classes.}
}
\value{
A distance matrix
}
\description{
Generate an adjusted distance matrix when the shift value \code{s} has already been selected.
}
