% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_test_dist_super.R
\name{aux_test_dist_super}
\alias{aux_test_dist_super}
\title{Supervised distance matrix adjustment for dimensionality reduction}
\usage{
aux_test_dist_super(
  mat = NULL,
  dmat = NULL,
  class.labels,
  s = NULL,
  nn.purity = NULL,
  k = 20,
  n = 50,
  ret.stat = FALSE
)
}
\arguments{
\item{mat}{A data frame or matrix to be processed. Rows are variables and columns are samples.}

\item{dmat}{A precomputed distance matrix. Used as an alternative input to \code{mat}.}

\item{class.labels}{A vector of class labels}

\item{s}{Shift value added to distances between samples from different classes.}

\item{nn.purity}{The purity of the nearest neighbors, 0~1}

\item{k}{Number of nearest neighbors}

\item{n}{Number of points used to interpolate the relationship between \code{s} and nearest-neighbor purity.}

\item{ret.stat}{Return the n point of the correspondence between "s" and purity. Choose FALSE to return the distance matrix}
}
\value{
A distance matrix or a data frame containing the correspondence between "s" and purity
}
\description{
Supervised distance matrix adjustment for dimensionality reduction
}
