% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_get_dist_super.R
\name{aux_get_dist_super}
\alias{aux_get_dist_super}
\title{Get supervised distance matrix with the 's' value}
\usage{
aux_get_dist_super(mat = NULL, dmat = NULL, class.labels, s = 10)
}
\arguments{
\item{mat}{A data frame or matrix that needs to be processed. row:variable col:sample}

\item{dmat}{If you have already calculated the distance matrix then enter it here}

\item{class.labels}{A vector of class labels}

\item{s}{If you have already calculated the "s" value then enter it here}
}
\value{
A distance matrix
}
\description{
Applicable when the appropriate s value has already been selected
}
