% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_test_dist_super.R
\name{aux_test_dist_super}
\alias{aux_test_dist_super}
\title{Data preprocessing for supervised t-SNE}
\usage{
aux_test_dist_super(
  mat = NULL,
  dmat = NULL,
  class.labels,
  s = NULL,
  nn.purity = NULL,
  k = 20,
  n = 50,
  ret.stat = FALSE
)
}
\arguments{
\item{mat}{A data frame or matrix that needs to be processed. row:variable col:sample}

\item{dmat}{Already calculated distance matrix}

\item{class.labels}{A vector of class labels}

\item{s}{Already calculated "s" value}

\item{nn.purity}{The purity of the nearest neighbors, 0~1}

\item{k}{Number of nearest neighbors}

\item{n}{Number of points to interpolate during liner regression}

\item{ret.stat}{Return the n point of the correspondence between "s" and purity. Choose FALSE to return the distance matrix}
}
\value{
A distance matrix or a data frame containing the correspondence between "s" and purity
}
\description{
Data preprocessing for supervised t-SNE
}
